64-bit version.

+++ src/crt1.c
@@ -8,10 +8,12 @@ extern int __libc_start_main(int (*main)(int, char**),
         void (*rtld_fini)(void),
         void* stack) __attribute__((__noreturn__));
 
-__attribute__((__externally_visible__, __section__(".text.startup._start"), __noreturn__))
+__attribute__((__externally_visible__, __section__(".text.startup._start"),
+            __noreturn__))
 int _start(void* stack) {
-    int argc=*(int*)stack;
-    char** argv=(void*)(&((int*)stack)[1]);
+    // TODO: _dl_fini etc.
+    int argc=*(size_t*)stack;
+    char** argv=(void*)(&((size_t*)stack)[1]);
 
     __libc_start_main(main, argc, argv, NULL, NULL, NULL, (void*)stack);
